list custom resource managed resources
