+++ cyclops-ui/src/utils/k8s/workload.tsx
@@ -1,0 +1,4 @@
+export interface Workload {
+  status: string;
+  pods: any[];
+}
